feat: changed module imports for types and interfaces
