+++ packages/connector/src/snap.ts
@@ -1,8 +1,8 @@
 import { SSISnapApi } from '@blockchain-lab-um/ssi-snap-types';
 import {
   saveVC,
-  getVCs,
-  getVP,
+  queryVCs,
+  createVP,
   togglePopups,
   changeInfuraToken,
   getDID,
@@ -36,8 +36,8 @@ export class MetaMaskSSISnap {
   public getSSISnapApi = async (): Promise<SSISnapApi> => {
     return {
       saveVC: saveVC.bind(this),
-      getVCs: getVCs.bind(this),
-      getVP: getVP.bind(this),
+      queryVCs: queryVCs.bind(this),
+      createVP: createVP.bind(this),
       togglePopups: togglePopups.bind(this),
       changeInfuraToken: changeInfuraToken.bind(this),
       getDID: getDID.bind(this),
