+++ packages/snap/src/rpc/vc/createVP.ts
@@ -1,7 +1,7 @@
 import { veramoCreateVP } from '../../utils/veramoUtils';
 import { VerifiablePresentation } from '@veramo/core';
 import { ApiParams } from '../../interfaces';
-import { AvailableVCStores } from 'src/constants';
+import { AvailableVCStores } from '@blockchain-lab-um/ssi-snap-types';
 
 type CreateVPRequestParams = {
   vcs: [
