+++ packages/snap/src/rpc/vcStore/getAvailableVCStores.ts
@@ -1,4 +1,4 @@
-import { availableVCStores } from '../../constants/index';
+import { availableVCStores } from '@blockchain-lab-um/ssi-snap-types';
 
 export function getAvailableVCStores(): string[] {
   return availableVCStores.map((key: string) => key);
