+++ packages/snap/src/rpc/vcStore/setVCStore.ts
@@ -1,4 +1,4 @@
-import { AvailableMethods, AvailableVCStores } from 'src/constants';
+import { AvailableMethods, AvailableVCStores } from '@blockchain-lab-um/ssi-snap-types';
 import { ApiParams } from '../../interfaces';
 import { snapConfirm } from '../../utils/snapUtils';
 import { updateSnapState } from '../../utils/stateUtils';
