+++ packages/snap/src/utils/keyPair.ts
@@ -6,7 +6,7 @@ import { ApiParams, SSISnapState } from 'src/interfaces';
 import { getAccountIndex, setAccountIndex } from './snapUtils';
 import { ethers } from 'ethers';
 import { _hexToUint8Array } from './snapUtils';
-import { didCoinTypeMappping } from '../constants/index';
+import { didCoinTypeMappping } from '@blockchain-lab-um/ssi-snap-types';
 import { SnapProvider } from '@metamask/snap-types';
 
 export async function getAddressKeyDeriver(
