+++ packages/snap/src/utils/veramoUtils.ts
@@ -1,5 +1,5 @@
 import { Agent, getAgent } from './../veramo/setup';
-import { VCQuery } from '@blockchain-lab-um/ssi-snap-types';
+import { VCQuery, AvailableVCStores } from '@blockchain-lab-um/ssi-snap-types';
 import {
   IIdentifier,
   MinimalImportableKey,
@@ -10,7 +10,6 @@ import {
 import { getCurrentDid } from './didUtils';
 import { getPublicKey, snapConfirm } from './snapUtils';
 import { SnapProvider } from '@metamask/snap-types';
-import { AvailableVCStores, availableVCStores } from '../constants/index';
 import { ApiParams } from '../interfaces';
 import { snapGetKeysFromAddress } from './keyPair';
 import { BIP44CoinTypeNode } from '@metamask/key-tree';
